requirements.txt
# LangChain and LangGraph
langchain
langchain-openai
langchain-pinecone # Online free VectorStore
langchain-community
langgraph
langchainhub

# Other stuff
python-dotenv

black

# PDF handling
pypdf # ingest PDFs
faiss-cpu # only need CPU, creates local vector store